test(ui): add ability to pass arbitrary state to renderWithRedux
